Add rake task for changing groups organisation
Sometimes groups need to move from on organisation to another.

Currently, this would have to be done through the rails console, or in
production through the DB.

This commit adds a new rake tasks to change the organisation of one or
more groups.

A new namespace, groups, has been added. This contains the tasks
change_organisation and change_organisation_dry_run.

This layout is supposed to match the other tasks we have.

To invoke the task from the command line, use:

```
rake groups:change_organisation_dry_run[VcyzrC3R,2]
```

The task takes one or more groups external_ids and must end with
the id of the target organisation.