Update forms in group move tasks
The non-bulk group move tasks weren't doing anything with the forms in a
moved group. This meant that these forms would end up in a different
organisation from the group they're in, which would be odd and may have
unintended consequences.

This commit changes it so that when a group moves, all of its forms also
have their organisation changed.

It's possible we'll get rid of/significantly change the organisation_id
field in the API soon - if this is the case then it may be possible to
delete most of this code and its tests.