Remove Form changes when changing group org
The rake task to move groups between organisations included updating
each form in the group to change the organisation_id held in the API.

Running the task in production failed when the Form did not have an
organisation set. This happened because trial groups don't have an
organisation_id set.

The task could either be changed to allow for Forms which don't have an
organisation_id, or we could not change the Forms themselves.

We no longer use the organisation_id field. The organisation of a form
is based only on the group it belongs to.

Because of this, I think its better to remove the code and tests, which
are simpler as a result.

This PR changes the task to not alter the Form organisation_id.