Remove organsations:summary rake task
We added this task when we were moving over to the new user management
model with groups. It's currently broken because the
`organisation.default_group` field no longer exists.

Rather than updating this task to deal with removing the
`form.organisation_id` field, we've decided to remove it. If we need
something similar in the future, we can add it back in.