Update test to reflect move to groups model
This test was checking that an error was thrown when the user does
not have permissions to view the form. However, it was still assuming
we were checking permissions based on the `organisation_id` of the
form, rather than whether the user is a member of the group the
form is in.

Update the test to reflect the behaviour it is now testing.