Update text max answers for repeatable question
This commit updates the text which shows when a form creator is making a
new question. Currently there is a warning that only 10 answers can be
given for a question which allows multiple answers.

That limit is being changed to 50. This commit updates the figure.

Further updates to the content are planned but this commit makes sure
the figure is correct for now.