Ensure single option seleciton questions can't be repeatable
In order to ensure that an invalid route can't be set up with a repeatable selection question, this change means that if a question is changed to a selection/only-one-option combination, it will be forced to not be repeatable. The option for changing one of these questions to repeatable will not be available either.