Fix name clash in tests
The tests for the support details were checking that the heading 'Email'
did not appear on the page in some circumstances. However, the previous
commit adds the heading 'Email' on a different part of the page. This
meant that the test was wrongly failing.

This commit updates the tests to use xpath so that the headings in
question are properly targeted and the tests don't collide. They're now
looking for the specific 'Email' heading after the relevant section
heading.