Remove custom linting plugin
This linting plugin didn't provide much value (users can run the
frontend lint in a separate terminal) and it slowed down the vite build.

This was especially annoying when starting bin/dev, since the main idea
behind Vite is to have a fast development server.

One consequence of removing this plugin is that vitest no longer
automatically runs the linting task when running the tests. We were
relying on this in the build_and_test github workflow, so this commit
also readds the linting task to the github action.