Add input object for share preview task
The view for the new task to share a preview of the form will use
the MarkCompleteComponent.

Add an input object that extends MarkCompleteInput, which is
responsible for passing the value of the radio button in the component
between the controller and view, and for performing the submit action
when the form is submitted.