Fix Sass 'mixed declarations' warning
This deprecation warning relates to the way that the sass declarations
are ordered. In v2, the way these declarations are turned into CSS will
change and if they're left as they are they'll generate more CSS than is
necessary.

There shouldn't be any visual differences as a result of these changes.

https://sass-lang.com/documentation/breaking-changes/mixed-decls/