order list of groups on groups index page
previously, these were shown in the order that records are returned from the database, which was not consistent - when a user refreshed the page the order might change.

I've added a default scope to the Group Application record to automatically reorder the groups to be in alphabetical order. Note that by default rails sorts by case first, then alphabetically, so I've added a lower case case to the order scope.