Improve have_been_updated RSpec matcher
This matcher was confusing, because it accepted the updated version of
the resource as the subject of the matcher. It makes more sense to
accept the version of the resource we expect to be updated as the
subject of the matcher, and the version of the resource we expect it
to be updated to as an argument.

Add a `have_been_updated_to` matcher that accepts the resource we
expect in the PUT request body as an argument. This is called like:

```
expect(form).to have_been_updated_to(updated_form)
```

Keep a matcher named `have_been_updated` that will check whether a PUT
request with any body has been made for the resource. The main use for
this is to use the negated matcher.

When we call `.not_to have_been_updated`, this expectaction should
fail if any PUT request was made for the resource. Previously, however,
this expecation passed if there was a PUT request made but it had a
different request body to that constructed from the subject of the
expectation.