Rename FormListService to FormListPresenter
Service Objects are an intermediary layer between "business objects" and "business logic" that exists outside of the model or controllers. Presenters are classes that is code only relevant to the View. In this case FormListServices was only fetching and formatting values to be output in the groups#show view, so isn't at all a Service.