Add share preview task to task list
Adds a link to the task list to take the user to the share preview page
added in 872af25.

This task will only be shown if the user can make the form live - in
other cases the whole make live section will be replaced by the message
telling the user they can't make the form live.

As part of this commit the tests have been rearranged a bit to better
reflect which behaviours are at task level and which are at section
level.