Only add one question for feature test with existing questions
We have 2 scenarios in this feature test, one to add questions when
there are no existing questions, and one to add questions when there
are existing questions.

In these 2 scenarios, the pages for the individual question types will
not differ. Therefore it is unneccessary to add all question types for
both scenarios.

For the scenario where there are existing questions, only add a
single question. This will speed up the running time for these tests.