Include share preview task in task count
Include the share preview task in the count of tasks that have been
completed at the top of the task list for a draft form. Do not include
it if the user does not have permission to make the form live, as they
will not be shown the task in this case.