Add sorting when presenting list of Forms
While we do not have a table for Forms in the database (hopefully this will not be the case for too long as we rework the API and port more objects concerning drafting into the Admin app) we cannot apply a default scope on the GroupForm class. So instead we do the sorting in the presenter, which feels appropriate as we mainly care at this point when the users views the list of their forms.