Add rake task to remove an empty group
When users create groups they don't want, we need a way to remove them
from the organisation and remove them. This can happen if a user selects
the wrong organisation at login.

This commit adds a simple rake task to remove a group when given a group
external id.

The group must not contain any forms. This avoids any complications
about what should happen to those forms.

To remove a group which has forms, remove the forms first.