Add report page for all forms with add another answer
This adds a new route to the reports controller, which leads to the full report for all forms that have a repeatable question. This report shows the form with a link to that form, and has a row for each question within the form that's currently set to be repeatable.