Add long_lists_enabled column to groups
This will be a temporary column to feature flag the ability to create
"select an option from a list" type questsions with more than 30
options, which will use an autocomplete component.

We initially want to restrict usage to one organisation to test with,
and having the flag on the group will allow us to do this.

We opted to do this over using the organisation level feature flagging
in the settings that we previously used as this had the following
issues:
- We could not enable for an organisation in dev/staging environments
without affecting production.
- For testing, we would want it enabled for the GDS organisation but
the GDS organisation has real forms and users.
- While this feature is only enabled for one organisation, if we
enabled it for all forms for our organisation, this would be
unrepresentative of most forms.