Update change links for selection question
When long_lists_enabled is true for the group, make the "Change" links
link to the pages in the new journey for creating a selection question

For the new journey, there are separate pages for choosing how many
options people can choose, and adding the options to select from. For
the old journey, these were on the same page.