Change how we set detault answer_settings for selection questions
We don't need to set default values for `only_one_option` or
`include_none_of_the_above` when configuring the default answer
settings for a selection question type. We only need to configure
default options so that the page renders as expected with 2 blank
fields for options.

We don't want to set these defaults for the new journey where
`only_one_option` and `include_none_of_the_above` are configured using
radio buttons, as we want the user to have to select an answer rather
than having one pre-selected.