Add page to choose type of selection question
Adds a new page to choose whether to allow people to pick only one
option or multiple options for a question to select from a list of
options.

This page will initially only been shown if the `long_lists_enabled`
flag is enabled on a group. Without this flag enabled, we offer this
option as a checkbox on the page to configure the settings for a
selection question.