Add page to add selection options
Adds a new page to provide the selection options for a "select from a
list of options" question.

This page will initially only been shown if the `long_lists_enabled`
flag is enabled on a group, after a page for the user to select
whether people can choose only one or multiple options.

This page differs from the existing page in that the option to select
whether people can choose only one or multiple options is now on a
separate page, and the question to choose whether to include an option
for "None of the above" is now radio buttons rather than a checkbox.