Update last_signed_in_at time for user whenever they authenticate
We want to be able to find out when users haven't signed in to our
service for a while. This commit will make it easier in future, by
starting to save the time when users sign in to the database.

I think that we don’t need the information to be that granular, so I
propose we update the time last_signed_in_at on the user record whenever
they authenticate. As our session policy requires form creators to sign
in every 20 hours and we're most likely interested in users who have
been inactive for 12 months or more this seems like a reasonable level
of resolution to me, and requires much fewer writes to the database.