Update maximum options for selection questions
On the new journey for adding a selection question, which the user
will be shown if `long_lists_enabled` is turned on for their group, we
will allow for users to add up to 1000 options, if they have selected
"People can choose only one option".

We display a warning if the user enters more than 1000 options, but
we do not display hint text on the page detailing the maximum.

When the user has selected "People can choose more than one option",
the maximum options will remain at 30 options.