Use 'match_array' to fix flaky Mailchimp test
Rspec provides a matcher 'match_array', which will match any array which
has only the elements specified. This is useful in cases where we want
to check that an array has the right items, but the order is
unimportant.

This Mailchimp rake task test was occasionally returning `false` because
the list of users to synchronise isn't always in the same order (I think
this is because they're returned in the default database order). This
commit uses match_array to remove that flakiness.