Add ConditionRepository to separate Condition calls
This adds a new ConditionRepository, a service which acts as the interface between the Condition model and the rest of the app. Rather than making direct calls to the model, other parts of the app will now go through the repository, which will make it easier for us to swap over to using locally stored conditions once the API data is brought across.

All tests that had stubbed out an API endpoint relating to Conditions have shifted over to stubbing out the ConditionRepository instead, and the appropriate endpoints are now tested within the repository.