Add helper for rendering time when user last signed in
We want to be able to show the date when a user last signed in.

However, if they have not signed in since the code to automatically
update the `last_signed_in_at` attribute was added in commit 87a9d73
then we won't know exactly when they last signed in, but we can infer a
date range based on whether they have signed in since Auth0 was enabled
in production (see alphagov/forms-deploy#401).

This commit adds a helper that makes it easy to display the right date
or date range, depending on the user.

Co-authored-by: hannahkc <hannah.cooper@digital.cabinet-office.gov.uk>