update seeds to add new form to end to end group
we're adding a new end to end test that will test forms with a submission type of s3.
There's no front end way to set a form  to a type of s3, so we can't create the form in the e2e tests.
Instead, when testing locally, we will use a new seeded form that has been created in forms-api.

This commit connects the new form and adds it to the end to end group.