Update content for number of selections on edit page
Update the content displayed in the summary table for the answer to
the question asking how many options people should be able to select
on the edit question page.

Keep the old content for the old journey when the `long_lists_enabled`
flag is turned off for the group.

Also update the content on the page to answer this question with the
latest iteration of content from the designs.