Change text for conditions in pagelist component
Conditions are displayed in the page list component with a text
description.

This commit changes the way conditions are described.

It:
- condenses the text into a single line paragraph
- shortens the content
- includes page numbers for the question a condition goes to

To support these changes, I've changed the tests. The tests now
explicitly check the result of getting a description against a string.

I chose this over referencing the translations because it's so much
easier to read and understand.

I chose to test the condition_description method over the testing the
sub-methods to try and test the behaviour rather than the implementation.

I considered checking the rendered component for the text. This makes
working out why a test failed difficult. Instead I've chosen to check
the result of the method and then check the text is included in the view.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>