Display a warning if number of options needs to be reduced
Display a warning on the "How many options should people be able to
select?" page if there are more than 30 existing selection options.

There is a limit of 30 options when the user selects
"People can select one or more options", so if the user is changing
their answer to this question we want to warn them that they will need
to reduce the number of options.

We already display a warning if the question has routes, as changing
to "People can select one or more options" will cause the routes to
be deleted. Reword this error to make sense with the new way we are
asking this question.

If we would display both warnings, insead display a single
notification banner with content about both these things.

Co-authored-by: Hannah Cooper <hannah.cooper@digital.cabinet-office.gov.uk>