Show error on edit question page when too many options
For a selection type question, when "People can select only one
option" is selected, up to 1000 options can be entered. When "People
can select one or more options" is selected, only 30 options can be
entered.

Add validation on the "Edit your question" page to ensure that the
form creator hasn't managed to switch the selection type to "People
can select one or more options" when there are more than 30 options.
This is possible if the form creator changes this setting and then
clicks the back button after they are taken to the page to edit their
selection options.

Display the error as an error in the summary list table, similar to
how we display routing errors on the list questions page.