Stop showing last signed in a year ago table if it is empty
The table of users who have `last_signed_in_at` < 1 year ago is currently
not useful, as the last_signed_in_at attribute was added less than a
year ago.

Originally this table had a message to show if it has no rows, as I was
thinking that in future it might be useful to have a placeholder, but
right now it just causes confusion. Let's get rid of the empty message
for now, and if we want it again in future we can revert this commit.