Fix duplication between tables in when user last signed in report
The "Person signed in between 3 November 2023 and 4 November 2024" table
and "Person signed in before 3 November 2023" table should be mutually
exclusive; users should appear in one table or the other but not both.

This commit adds a test checking this, and fixes the database query so
that this is the case.