Change titles for secondary skips in page list
Routes should be numbered in the page list according to the check_page
they belong to.

Adding this to the page list without making multiple calls to the API in
a clean way is tricky.

This commit uses a memoized methog, routing_conditions_with_index, which
containst a hash of page_id => [condition, index] pairs. This is then
used to generate list of conditions for individual pages.