Add a PageRepository and reroute Page model usage through it
The PageRepository being introduced is used as an interface between the application and any Page method calls that result in an API call through ActiveResource. The intention is to create a layer of separation, so that database requests can be rerouted more easily as part of the Admin/API responsibilities shifting.