Add username and password auth for user research environment
We want to stop using HTTP Basic Auth in the user research environment,
because it seems to work less and less with the browsers of our
research participants. We still don't want to be using the full-fat
Auth0 authentication though, as we want to be able to set up test user
accounts in advance and stop the general public from using this
environment.

This commit adds a HTML form for signing in, with the same level of
security as the HTTP Basic Auth we were previously using, but more
likely to be compatible with modern browsers.

We use OmniAuth to create the form and compare the credentials, which is
a bit of an awkward fit, but is simple to start with. If the credentials
don't match, we just bounce the user out to our even less pretty auth
failure page, with a link to retry. None of this is using the GOV.UK
Design System styles and patterns like the rest of the site, but again
I think this is fine for now.