Add secondary_skip_input
Allow users to add a second skip condition to enable two-branch routing.

This commit adds a new input model for skip conditions.

It's based largely on the condition input model. It currently only
supports creating new ones, not changing existing records, which will be
added later.

The model includes helpers which will be useful for the view which will
render the form. These can probably be refactored into other classes.

The validations test the basic functionality listed in the design. There
are no validations for ensuring the routing_page_id and goto_page_id are
valid ids. I think we need those.

Co-authored-by: Hannah Content <hannah.cooper@digital.cabinet-office.gov.uk>