Allow auth provider string to contain underscore
If we have an auth provider as a symbol we want it to be word separated
with underscores, but in the URL we want it to be word separated by
dashes.

This does mean that the OmniAuth strategy needs to have the request and
callback paths set to use the provider name with dashes instead of
underscores.

Also removing the trailing slash, as it is unneeded.