Refactor question input to use create! for new Page
This relocates the Page creation from the controller, and puts it inside the QuestionInput object. This means that all the page validation and creation is handled within the object, rather than being split with the controller.

Also refactors the question input so that creating and updating are handled by separate methods. This wasn't needed originally, as the save! action was covering both cases.