Log entry method for selection options
We want to know how many people are using the text box or individual
fields to enter their options when creating a "Select from a list of
options" question. We want to know whether this depends on how many
options they are entering.

Log when options are submitted, with:
- The input method (is_bulk_input)
- The number of options (options_count)
- Whether people can select more than one option (only_one_option)

We will use this log line to create a report in Splunk