Fix update to existing secondary_skips
When updating existing skip routes, they were always being deleted and
recreated.

This happened because routing_page_id in the model is taken from the
params and is a string. When taken from a Condition, it's an integer.

This commit fixes this in the most basic way by ensuring that both ids
are compared as strings.

There are other options:
- using ActiveModel attributes to specify routing_page_id is always
  converted to an integer
- never updating existing conditions but always removing and creating a
  new one

The first option is neat and prevents the issue happening elsewhere in
the code. I've held off using it because goto_page_id can't be treated
as an integer as it can take "check_your_answers" as well as a page id.

I think a solution to both would be nicer.

It's possible that we should always delete and create a new condition
when the secondary skip is updated. I think that's something worth
discussing but feels like a bigger decision which isn't tackled here.

Either way, this issue is now covered by a test and is fixed for now.