Fix secondary_skip#update
When updating an existing secondary skip, the condition was being
updated correctly with both routing_page_id and goto_page_id.

This was because assign_values was mistakenly beiing called in the
update action.

This call overwrites the model values with values from the existing
Conndition, which meant that the conditions would not be updated!

This commit removes the call to assign_values as it should not be used
here.