Update content on page to add a question route
With the introduction of branch routing we want to update the content on
the page to add a route from a question.

This commit updates the content both when the branch routing feature is
enabled and when it is not.

The first paragraph which explains routing has been changed. When the
branch routing feature is enabled, then there is an additional paragraph
explaining how branch routes can be created.

We have also updated the page title.

Co-authored-by: hannahkc <hannah.cooper@digital.cabinet-office.gov.uk>
Co-authored-by: charlotteharry <122444809+C-Harry@users.noreply.github.com>