Fix pagelist component's display of secondary_skip
This commit fixes how secondary_skip's are displayed in the pagelist.

Currently the text of the check_page is used. The check_page in a
secondary skip question is the radio question.

The correct text is the routing_page. The routing_page is the question
which the condition will jump from.

This commit updates the text used and renames the methods and locale
entries to reflect the change.