Add delete link to summary card for any other answers route
Add a link to the actions for the any other answers route summary card
that will lead to a page to confirm deletion of the route condition.

Currently the link goes nowhere, but we want to add it now so we can
put it in front of users in testing and see how well it works.

This link should not appear if the branch routing feature flag is not
enabled.