Add requires condition shared_example
Each action with the secondary_skip controller relies on the page
already having a condition. This is enforced using a before_action
within the controller.

This was tested in the specs by repeatedly setting up the same pages and
running the same expectation with a different request.

This commit extracts that check to a shared example to make the intent
clearer and reduce duplication.