SPIKE: Delete users with no name/organisation
This tests deleting users that have not set their organisation or name
along with deleting their forms.

Users are skipped if they are a member of a group - as we do not
expect any users who have not set their organisation or name to be a
member of a group, as groups were introduced after we started asking
for the organisation and name, and we only put people's forms in a
default group once they have entered their organisation.

If users do have a group, it probably makes sense to double check that
they are the only member of the group and run the dedicated task to
remove that group.

We delete the user's account in auth0 - which requires the
"Auth0 Management API" to be enabled for the application in the Auth0
tenant with the `delete:users` permission enabled.