Fix error in routes#show for missing goto_page
When a user has created a "basic route" and then deletes the page the
condition points to with it's "goto_page_id", the show route throws an
exception when the question_text cannot be accessed.

This commit fixes this error by using placeholder text when the goto_page
cannot be found.

This is a short term fix intended to stop an error being raised, rather
than a considered design fix.