Stop form creators from making repeatable checkbox questions
We don't think it makes sense for a selection question to allow
multiple answers. In PR alphagov/forms-admin#1453 we hid the option for
selection questions where only one option is allowed, but we should have
done it for all kinds of selection questions.

This commit changes the edit question page so that selection questions
can't be repeatable.