Fix DefaultGroupService to handle trial users with the same name as another
In PR alphagov/forms-admin#1236 we added DefaultGroupService to help
with migrating users with trial accounts who had not set a name or
organisation but had created forms to standard accounts with groups.
The `DefaultGroupService#create_user_default_trial_group!` method
creates a group in the chosen organisation of the user named after the
chosen name of the user.

At around the same time in PR alphagov/forms-admin#1206 we added a
constraint that more than one group could not have exist with the same
organisation and name.

At the time we did not account for the possibility that the default
group service would interact with this constraint, partly because the
set of users who would be affected is small. However, we have now had an
incident where two users who have two user accounts each have run into
this issue; see the incident report [[1]] and Sentry issue [[2]] for
more details.

To fix this we've added logic to append an incrementing number to the
user's name in the name of the group to be created. This will work for
the users who are currently unable to sign in due to this issue, and any
new cases. We still consider it unlikely to happen again, but at least
now we're covered if it does.

[1]: https://docs.google.com/document/d/1BZhgaNGaeV4dyZGkEmmDywoceEEegtxzjDkmULc5Npk/edit
[2]: https://govuk-forms.sentry.io/issues/6089472300

Co-authored-by: Hannah Content <hannah.cooper@digital.cabinet-office.gov.uk>
Co-authored-by: Kelvin Gan <dev@kelvingan.co.uk>