Change DefaultGroupService to use user's existing active group if exists
In the previous commit we added handling for the case when creating a
default trial group for a user where a group exists with the same name
but different creator; this also inadvertently affected the case where
a group already exists with the same name and same creator but
different status.

As we don't want to put forms created when the user had a trial account
into an active group where those forms can be made live, and as it
should not really be possible for a user to have a group before setting
their organisation, previously we just handled this by allowing the
ActiveRecord::RecordInvalid exception to bubble up due to the duplicate
name, but this no longer happens.

This commit fixes the failing test by explicitly re-raising the
exception if a default trial group exists for the user but which is
active.