Remove 'long_lists' from variable names
This test file uses variables to store the paths it's looking for. We
originally named these with `long_lists` to avoid a name collision with
the old selection routes. Those old routes no longer exist, so we can
rename the variables to make them simpler and more consistent with the
route names.