Remove plural in 'with_selections_settings` trait
The `with_selections_settings` trait was using the plural `selections`.
This was consistent with the old selection logic, but inconsistent with
both the other settings traits (e.g. `with_date_settings`) and the new
selection logic (which uses the singular `selection`).

This commit renames all instances of `with_selections_settings` to the
singular `with_selection_settings`.