Add delete actions to secondary skip controller
This adds delete and destroy actions to the secondary skip controller. These are used to remove a secondary skip from a question's routing, to complete the full CRUD functionality.

A separate delete_secondary_skip_input object has also been created, to handle the validations and submission for the delete page.