Remove unnecessary call to Page.new from questions controller
We want to abstract away as many calls to the Page model as possible
using the page repository added in PR #1587.

This call to Page.new is unnecessary, as the view does not use it (see
previous commits), so we can remove it, rather than having to worry
about adding a new method to the repository.